+++ packages/cli/src/export.test.ts
@@ -1,6 +1,8 @@
 import { describe, it } from "@jest/globals";
 import { hi18n } from "./command";
-import { MockedOutput, withProject } from "./testing/common";
+import { MockedOutput, initFixtures } from "@hi18n/dev-utils";
+
+const { withProject } = initFixtures(__dirname);
 
 describe("export", () => {
   // eslint-disable-next-line jest/expect-expect
