+++ packages/connector-i18n-js/.eslintrc.js
