+++ packages/connector-i18n-js/.prettierignore
@@ -1,0 +1,2 @@
+/dist
+/src/__fixtures__/*/expect-*
