+++ packages/connector-i18n-js/.prettierrc.json
@@ -1,0 +1,1 @@
+{}
