+++ packages/connector-i18n-js/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 0.1.0
+
+Initial release.
