+++ packages/connector-i18n-js/configs/tsconfig.main.json
