+++ packages/connector-i18n-js/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@hi18n/cli",
-  "version": "0.1.9",
-  "description": "Message internationalization meets immutability and type-safety - command line tool",
+  "name": "@hi18n/connector-i18n-js",
+  "version": "0.1.0",
+  "description": "Converter between i18n-js and @hi18n/core",
   "keywords": [
     "intl",
     "i18n",
@@ -22,7 +22,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/wantedly/hi18n.git",
-    "directory": "packages/cli"
+    "directory": "packages/connector-i18n-js"
   },
   "author": "Masaki Hara <ackie.h.gmai@gmail.com>",
   "license": "MIT",
@@ -40,14 +40,8 @@
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    },
-    "./api": {
-      "types": "./dist/api.d.ts",
-      "default": "./dist/api.js"
-    }
+    "types": "./dist/index.d.ts",
+    "default": "./dist/index.js"
   },
   "scripts": {
     "babel": "babel -x .ts src -d dist --source-maps --ignore '**/*.d.ts,**/*.test.ts,src/__fixtures__/**/*,src/testing/**/*'",
@@ -61,35 +55,22 @@
     "tsc": "tsc --build"
   },
   "dependencies": {
-    "@babel/core": "^7.17.10",
-    "@babel/eslint-parser": "^7.17.0",
     "@babel/runtime": "^7.18.3",
-    "@hi18n/eslint-plugin": "workspace:^0.1.8",
     "@hi18n/tools-core": "workspace:^0.1.0",
-    "@typescript-eslint/utils": "^5.21.0",
-    "commander": "^9.2.0",
-    "core-js-pure": "^3.23.2",
-    "cosmiconfig": "^7.0.1",
-    "eslint": "^8.13.0",
-    "glob": "^8.0.1",
-    "resolve": "^1.22.0"
+    "core-js-pure": "^3.23.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.18.9",
+    "@babel/core": "^7.17.10",
     "@babel/plugin-transform-runtime": "^7.18.9",
     "@babel/preset-env": "^7.18.9",
     "@babel/preset-typescript": "^7.18.6",
-    "@babel/types": "^7.18.9",
-    "@hi18n/core": "workspace:*",
     "@hi18n/dev-utils": "workspace:*",
     "@jest/globals": "^28.1.3",
-    "@types/eslint": "^8.4.5",
-    "@types/glob": "^7.2.0",
-    "@types/resolve": "^1.20.2",
     "babel-jest": "^28.1.3",
     "babel-plugin-polyfill-corejs3": "^0.5.3",
+    "eslint": "^8.13.0",
     "eslint-plugin-node": "^11.1.0",
-    "expect": "^28.1.3",
     "jest": "^28.1.3",
     "prettier": "^2.7.1",
     "rimraf": "^3.0.2",
