+++ packages/connector-i18n-js/src/index.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, it } from "@jest/globals";
+
+describe("something", () => {
+  it("does something useful", () => {
+    expect(1 + 1).toBe(2);
+  });
+});
