+++ packages/connector-i18n-js/src/index.ts
@@ -1,0 +1,1 @@
+import { Connector } from "@hi18n/tools-core";
