+++ packages/dev-utils/.prettierrc.json
@@ -1,0 +1,1 @@
+{}
