+++ packages/dev-utils/configs/tsconfig.base.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../configs/tsconfig.base.json",
+  "compilerOptions": {
+    "lib": ["ESNext"],
+    "module": "commonjs"
+  }
+}
