+++ packages/dev-utils/configs/tsconfig.main.json
@@ -7,5 +7,5 @@
     "tsBuildInfoFile": "../dist/tsconfig.main.tsbuildinfo"
   },
   "include": ["../src/**/*.ts"],
-  "exclude": ["../**/*.test.ts", "../src/__fixtures__/**/*.ts"]
+  "exclude": ["../**/*.test.ts"]
 }
