+++ packages/dev-utils/src/index.test.ts
@@ -1,0 +1,5 @@
+import { expect, test } from "@jest/globals";
+
+test("something", () => {
+  expect(42).toBe(42);
+});
