+++ packages/tools-core/.prettierrc.json
@@ -1,0 +1,1 @@
+{}
